moved AwcInfrastructureAggregationHelper to own file
